+++ src/components/dialog/dialog.tsx
@@ -246,7 +246,12 @@ export class Dialog {
                   {this.label || String.fromCharCode(65279)}
                 </slot>
               </span>
-              <sl-icon-button part="close-button" class="dialog__close" name="x" onClick={this.handleCloseClick} />
+              <sl-icon-button
+                exportparts="base:close-button"
+                class="dialog__close"
+                name="x"
+                onClick={this.handleCloseClick}
+              />
             </header>
           )}